split metric from Metrics Client
